requirements.txt
fastf1
pandas
python-dotenv